extract GlobalSearch into own component
