+++ packages/theme/components/layout/global-search/GlobalSearch.module.css
@@ -1,0 +1,20 @@
+.GlobalSearch__searchInput {
+  width: 100%;
+}
+
+.GlobalSearch__searchResultsList {
+  list-style: none;
+  padding: 0;
+  border-top: var(--brand-borderWidth-thin) solid var(--brand-color-border-default);
+  padding-block-start: var(--base-size-16);
+}
+
+.GlobalSearch__searchResultsList li:not(:last-child) {
+  margin-block-end: var(--base-size-16);
+  padding-block-end: var(--base-size-16);
+  border-bottom: var(--brand-borderWidth-thin) solid var(--brand-color-border-default);
+}
+
+.GlobalSearch__searchResultsList a {
+  color: var(--brand-color-text-default);
+}
