+++ packages/theme/components/layout/header/Header.module.css
@@ -71,10 +71,6 @@
   }
 }
 
-.Header__searchInput {
-  width: 100%;
-}
-
 .Header__siteTitle {
   display: flex;
   gap: var(--base-size-8);
@@ -85,21 +81,4 @@
 
 .Header__siteTitle svg {
   fill: var(--brand-color-text-default);
-}
-
-.Header__searchResultsList {
-  list-style: none;
-  padding: 0;
-  border-top: var(--brand-borderWidth-thin) solid var(--brand-color-border-default);
-  padding-block-start: var(--base-size-16);
-}
-
-.Header__searchResultsList li:not(:last-child) {
-  margin-block-end: var(--base-size-16);
-  padding-block-end: var(--base-size-16);
-  border-bottom: var(--brand-borderWidth-thin) solid var(--brand-color-border-default);
-}
-
-.Header__searchResultsList a {
-  color: var(--brand-color-text-default);
 }
