+++ packages/theme/components/layout/header/Header.tsx
@@ -1,10 +1,9 @@
-import React, {useCallback, useEffect, useMemo, useRef, useState} from 'react'
+import React, {useCallback, useEffect, useState} from 'react'
 import {MarkGithubIcon, MoonIcon, SearchIcon, SunIcon, ThreeBarsIcon, XIcon} from '@primer/octicons-react'
-import {Box, FormControl, IconButton, TextInput} from '@primer/react'
-import {Heading, Stack, Text} from '@primer/react-brand'
+import {Box, IconButton} from '@primer/react'
+import {Stack, Text} from '@primer/react-brand'
 import {clsx} from 'clsx'
-import {MdxFile, PageMapItem} from 'nextra'
-import {debounce} from 'lodash'
+import {PageMapItem} from 'nextra'
 
 import Link from 'next/link'
 import styles from './Header.module.css'
@@ -12,6 +11,7 @@ import {NavDrawer} from '../nav-drawer/NavDrawer'
 import {useNavDrawerState} from '../nav-drawer/useNavDrawerState'
 import {useColorMode} from '../../context/color-modes/useColorMode'
 import {DocsItem} from '../../../types'
+import {GlobalSearch} from '../global-search/GlobalSearch'
 
 type HeaderProps = {
   pageMap: PageMapItem[]
@@ -19,143 +19,14 @@ type HeaderProps = {
   siteTitle: string
 }
 
-type SearchResults = {
-  title: string
-  description: string
-  url: string
-}
-
 export function Header({pageMap, siteTitle, flatDocsDirectories}: HeaderProps) {
   const {colorMode, setColorMode} = useColorMode()
-  const inputRef = useRef<HTMLInputElement | null>(null)
-  const searchResultsRef = useRef<HTMLElement | null>(null)
   const [isNavDrawerOpen, setIsNavDrawerOpen] = useNavDrawerState('768')
   const [isSearchOpen, setIsSearchOpen] = useState(false)
-  const [isSearchResultOpen, setIsSearchResultOpen] = useState(false)
-  const [searchResults, setSearchResults] = useState<SearchResults[] | undefined>()
-  const [searchTerm, setSearchTerm] = useState<string | undefined>('')
-  const [activeDescendant] = useState<number>(-1)
-
-  useEffect(() => {
-    if (isSearchOpen && inputRef.current) {
-      inputRef.current.focus()
-    }
-  }, [isSearchOpen])
-
-  useEffect(() => {
-    const handleKeyDown = (event: KeyboardEvent) => {
-      if (event.key === 'Escape') {
-        setIsSearchResultOpen(false)
-      }
-    }
-
-    const handleClickAway = (event: MouseEvent) => {
-      if (
-        inputRef.current &&
-        !inputRef.current.contains(event.target as Node) &&
-        !searchResultsRef.current?.contains(event.target as Node)
-      ) {
-        setIsSearchResultOpen(false)
-      }
-    }
-
-    document.addEventListener('keydown', handleKeyDown)
-    document.addEventListener('click', handleClickAway)
-
-    // Clean up the event listener when the component unmounts
-    return () => {
-      document.removeEventListener('keydown', handleKeyDown)
-      document.removeEventListener('click', handleClickAway)
-    }
-  }, [])
 
   useEffect(() => {
     document.documentElement.setAttribute('data-color-mode', colorMode)
   }, [colorMode])
-
-  const setSearchResultsDebounced = debounce((data: SearchResults[] | undefined) => setSearchResults(data), 1000)
-  const searchData = useMemo(
-    () =>
-      flatDocsDirectories
-        .map(item => {
-          if (item.route === '/') return null // remove homepage
-          return item
-        })
-        .filter(Boolean)
-        .map(item => {
-          const {frontMatter, route} = item as MdxFile
-          if (!frontMatter) return null
-          const result = {
-            title:
-              frontMatter['show-tabs'] && frontMatter['tab-label']
-                ? `${frontMatter.title} | ${frontMatter['tab-label']}`
-                : frontMatter.title
-                  ? frontMatter.title
-                  : '',
-            description: frontMatter.description ? frontMatter.description : '',
-            url: route,
-          }
-          return result
-        }),
-    [flatDocsDirectories],
-  )
-
-  const handleChange = () => {
-    if (!inputRef.current) return
-    if (inputRef.current.value.length === 0) {
-      setSearchTerm(undefined)
-      setSearchResults(undefined)
-      setIsSearchResultOpen(false)
-      return
-    }
-    if (inputRef.current.value.length > 2) {
-      const curSearchTerm = inputRef.current.value.toLowerCase()
-
-      // filters the frontMatter title and descriptions against the search term
-      const filteredData = searchData
-        .filter((data): data is SearchResults => data !== null)
-        .filter(data => {
-          if (!data.title) return false
-          const title = data.title.toLowerCase()
-          const description = data.description.toLowerCase()
-          return title.includes(curSearchTerm) || description.includes(curSearchTerm)
-        })
-
-      // sorts the data to show hits in title first, description second
-      const sortedData = filteredData.sort((a, b) => {
-        const aTitle = a.title.toLowerCase()
-        const bTitle = b.title.toLowerCase()
-        const aIncludes = aTitle.includes(curSearchTerm)
-        const bIncludes = bTitle.includes(curSearchTerm)
-
-        if (aIncludes && !bIncludes) {
-          return -1
-        } else if (!aIncludes && bIncludes) {
-          return 1
-        } else {
-          return 0
-        }
-      })
-
-      setSearchResultsDebounced(sortedData)
-
-      setSearchTerm(inputRef.current.value)
-      setIsSearchResultOpen(true)
-      return
-    }
-  }
-
-  const handleSubmit = (e: React.FormEvent) => {
-    e.preventDefault()
-    if (!inputRef.current) return
-    if (!inputRef.current.value) {
-      // eslint-disable-next-line i18n-text/no-en
-      alert(`Enter a value and try again.`)
-      return
-    }
-
-    alert(`Name: ${inputRef.current.value}`)
-  }
 
   const handleSearchButtonOpenClick = useCallback(() => {
     setIsSearchOpen(true)
@@ -174,110 +45,7 @@ export function Header({pageMap, siteTitle, flatDocsDirectories}: HeaderProps) {
         </Text>
       </Link>
       <div className={clsx(styles.Header__searchArea, isSearchOpen && styles['Header__searchArea--open'])}>
-        <FormControl>
-          <FormControl.Label visuallyHidden>Search</FormControl.Label>
-          <TextInput
-            contrast
-            className={styles.Header__searchInput}
-            leadingVisual={<SearchIcon />}
-            placeholder={`Search ${siteTitle}`}
-            ref={inputRef}
-            onSubmit={handleSubmit}
-            onChange={handleChange}
-            trailingAction={
-              searchTerm ? (
-                <TextInput.Action
-                  onClick={() => {
-                    if (inputRef.current) {
-                      inputRef.current.value = ''
-                      setSearchTerm(undefined)
-                      setSearchResults(undefined)
-                    }
-                  }}
-                  icon={XIcon}
-                  aria-label="Clear input"
-                  tooltipDirection="nw"
-                  sx={{color: 'fg.subtle'}}
-                />
-              ) : undefined
-            }
-            aria-activedescendant={activeDescendant === -1 ? undefined : `search-result-${activeDescendant}`}
-          />
-        </FormControl>
-        {searchTerm && (
-          <Box
-            ref={searchResultsRef}
-            sx={{
-              display: isSearchResultOpen ? 'block' : 'none',
-              marginTop: 1,
-              position: 'absolute',
-              zIndex: 1,
-              backgroundColor: 'var(--brand-color-canvas-default)',
-              padding: 'var(--base-size-16)',
-              width: '100%',
-              maxWidth: ['calc(100% - 46px)', null, '350px'],
-              border: 'var(--brand-borderWidth-thin) solid var(--brand-color-border-default)',
-              borderRadius: 'var(--brand-borderRadius-medium)',
-              maxHeight: 300,
-              overflowY: 'auto',
-              overflowX: 'hidden',
-              '&::-webkit-scrollbar': {
-                width: 8,
-              },
-              '&::-webkit-scrollbar-track': {
-                backgroundColor: 'var(--brand-color-canvas-default)',
-              },
-              '&::-webkit-scrollbar-thumb': {
-                backgroundColor: 'var(--brand-color-text-muted)',
-                borderRadius: 'var(--base-size-4)',
-              },
-            }}
-          >
-            <Stack direction="vertical" padding="none" gap="none">
-              {searchTerm && (
-                <Box sx={{pl: 1}}>
-                  <Heading as="h3" size="subhead-large" id="search-results-heading">
-                    {searchResults && searchResults.length} Results for &quot;{searchTerm}&quot;
-                  </Heading>
-                </Box>
-              )}
-              {searchResults && searchResults.length > 0 ? (
-                <ul
-                  role="listbox"
-                  tabIndex={0}
-                  aria-labelledby="search-results-heading"
-                  className={clsx(styles.Header__searchResultsList)}
-                >
-                  {searchResults.map((result, index) => (
-                    <li
-                      key={`${result.title}-${index}`}
-                      id={`search-result-${index}`}
-                      role="option"
-                      aria-selected={index === activeDescendant}
-                    >
-                      <Text size="200" className={styles.Header__searchResultItemTitle}>
-                        <Link href={result.url}>{result.title}</Link>
-                      </Text>
-                      <Text
-                        as="p"
-                        size="100"
-                        variant="muted"
-                        id={`search-result-item-desc${index}`}
-                        className={styles.Header__searchResultItemDesc}
-                      >
-                        {result.description}
-                      </Text>
-                    </li>
-                  ))}
-                </ul>
-              ) : (
-                <Box sx={{p: '100', display: 'flex', justifyContent: 'center', alignItems: 'center', height: 150}}>
-                  <Text variant="muted">No results found</Text>
-                </Box>
-              )}
-            </Stack>
-          </Box>
-        )}
+        <GlobalSearch pageMap={pageMap} siteTitle={siteTitle} flatDocsDirectories={flatDocsDirectories} />
         <div className={styles.Header__searchHeaderBanner}>
           <Stack direction="horizontal" padding="none" gap={4} alignItems="center" justifyContent="space-between">
             <Text as="p" size="300" weight="semibold">
